LayoutTests/http/tests/resourceLoadStatistics/add-partitioning-to-redirect-expected.txt

Tests that partitioning is added mid-flight in redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame2-->'
--------
Redirect case, should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame3-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame4-->'
--------
Redirect case, should receive no cookie.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame5-->'
--------
Set partitioned, third-party cookie.


--------
Frame: '<!--frame6-->'
--------
Should receive partitioned cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'partitionedCookie'.
Client-side document.cookie: partitionedCookie=value

--------
Frame: '<!--frame7-->'
--------
Redirect case, should receive partitioned cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'partitionedCookie'.
Client-side document.cookie: partitionedCookie=value
